Organized some container
better structure of the containers folder and the resulting resources for each
