+++ front-end/src/components/Researcher/Launch/BountyLaunch.js
